pr-sug: move integration tests to './Serilog/Sinks/ApplicationInsights'
